feat: 이미지 엔티티 삭제
- 게시글과 이미지 간의 일대다 관계 제거
- 게시글에 이미지 데이터를 직접 통합
- 이미지 관련 DTO, 서비스, 레포지토리, API 컨트롤러 삭제
- 예시 이미지 4 추가
