+++ Docker/usedStuff/src/main/java/pposonggil/usedStuff/api/Board/BoardApiController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import pposonggil.usedStuff.dto.Board.BoardDto;
 import pposonggil.usedStuff.service.Board.BoardService;
 
@@ -18,6 +19,7 @@ public class BoardApiController {
 
     /**
      * 전체 게시글 조회
+     *
      * @return 게시글 Dto 리스트
      */
     @GetMapping("/api/boards")
@@ -27,6 +29,7 @@ public class BoardApiController {
 
     /**
      * 특정 게시글 상세 조회
+     *
      * @param boardId : 조회할 게시글 아이디
      * @return 게시글 아이디로 조회한 게시글 Dto
      */
@@ -37,33 +40,37 @@ public class BoardApiController {
 
     /**
      * 작성자 아이디로 게시글 조회
+     *
      * @param memberId : 조회할 회원 아이디
      * @return 회원 아이디로 조회한 게시글 Dto 리스트
      */
     @GetMapping("/api/boards/by-member/{memberId}")
-    public List<BoardDto> findImageBoardsByWriterId(@PathVariable Long memberId) {
-        return boardService.findImageBoardsByWriterId(memberId);
+    public List<BoardDto> findBoardsByWriterId(@PathVariable Long memberId) {
+        return boardService.findBoardsByWriterId(memberId);
     }
 
     /**
-     * 작성자 & 이미지 & 게시글 조회
-     * @return 이미지 게시글 Dto 리스트
+     * 작성자 & 게시글 조회
+     *
+     * @return 게시글 Dto 리스트
      */
-    @GetMapping("/api/boards/with-member-image")
+    @GetMapping("/api/boards/with-member")
     public List<BoardDto> getBoardsWithMember() {
         return boardService.findAllWithMember();
     }
 
     /**
      * 게시글 작성
+     *
      * @param boardDto : 작성할 내용이 담긴 게시글 Dto
      * @return 성공 -->
-     *          "boardId" : [Id]
-     *          "message" : "게시글을 작성을 완료했습니다."
+     * "boardId" : [Id]
+     * "message" : "게시글을 작성을 완료했습니다."
      */
     @PostMapping("/api/board")
-    public ResponseEntity<Object> createBoard(@RequestBody BoardDto boardDto) {
-        Long boardId = boardService.createBoard(boardDto);
+    public ResponseEntity<Object> createBoard(@RequestPart("boardDto") BoardDto boardDto,
+                                              @RequestPart(value = "file", required = false) MultipartFile file) throws Exception {
+        Long boardId = boardService.createBoard(boardDto, file);
 
         Map<String, Object> response = new HashMap<>();
         response.put("boardId", boardId);
@@ -73,19 +80,23 @@ public class BoardApiController {
 
     /**
      * 게시글 수정
-     * @param boardId : 게시글 아이디
+     *
+     * @param boardId  : 게시글 아이디
      * @param boardDto : 수정한 게시글 Dto
+     * @param file     : 수정할 사진 파일 (선택 사항)
      * @return 성공 -->
-     *          "boardId" : [Id]
-     *          "message" : "게시글을 수정을 완료했습니다."
+     * "boardId" : [Id]
+     * "message" : "게시글을 수정을 완료했습니다."
      */
     @PutMapping("/api/board/{boardId}")
-    public ResponseEntity<Object> updateBoard(@PathVariable Long boardId, @RequestBody BoardDto boardDto) {
+    public ResponseEntity<Object> updateBoard(@PathVariable Long boardId,
+                                              @RequestPart("boardDto") BoardDto boardDto,
+                                              @RequestPart(value = "file", required = false) MultipartFile file) throws Exception {
         BoardDto oldBoardDto = boardService.findOne(boardId);
         if (oldBoardDto == null) {
             return ResponseEntity.notFound().build();
         }
-        boardService.updateBoard(boardDto);
+        boardService.updateBoard(boardDto, file);
 
         Map<String, Object> response = new HashMap<>();
         response.put("boardId", boardId);
@@ -95,6 +106,7 @@ public class BoardApiController {
 
     /**
      * 게시글 삭제
+     *
      * @param boardId : 게시글 아이디
      * @return 성공 --> "게시글을 삭제하였습니다."
      */
