+++ Docker/usedStuff/src/main/java/pposonggil/usedStuff/repository/image/custom/CustomImageRepository.java
@@ -1,11 +1,11 @@
-package pposonggil.usedStuff.repository.image.custom;
-
-import pposonggil.usedStuff.domain.Image;
-
-import java.util.List;
-
-public interface CustomImageRepository {
-    List<Image> findAllWithBoard();
-    List<Image> findImagesByImageBoardId(Long boardId);
-
-}
+//package pposonggil.usedStuff.repository.image.custom;
+//
+//import pposonggil.usedStuff.domain.Image;
+//
+//import java.util.List;
+//
+//public interface CustomImageRepository {
+//    List<Image> findAllWithBoard();
+//    List<Image> findImagesByImageBoardId(Long boardId);
+//
+//}
