+++ Docker/usedStuff/src/test/java/pposonggil/usedStuff/service/ChatRoomServiceTest.java
@@ -40,7 +40,7 @@ class ChatRoomServiceTest {
     private Long chatRoomId1, chatRoomId2, chatRoomId3;
 
     @BeforeEach
-    public void setUp() {
+    public void setUp() throws Exception {
         // 회원 1, 2, 3생성
         memberId1 = createMember("name1", "nickName1", "01011111111");
         memberId2 = createMember("name2", "nickName2", "01022222222");
@@ -158,7 +158,7 @@ class ChatRoomServiceTest {
 
 
     public Long createBoard(Long savedId, String title, String content, LocalDateTime startTime, LocalDateTime endTime,
-                            TransactionAddress address, Long price, boolean isFreebie) {
+                            TransactionAddress address, Long price, boolean isFreebie) throws Exception {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd-HH:mm");
         String formatStartTime = startTime.format(formatter);
         String formatEndTime = endTime.format(formatter);
