+++ Docker/usedStuff/src/test/java/pposonggil/usedStuff/service/ReviewServiceTest.java
@@ -48,7 +48,7 @@ class ReviewServiceTest {
     private Long reviewId1, reviewId2, reviewId3, reviewId4;
 
     @BeforeEach
-    public void setUp() {
+    public void setUp() throws Exception {
         // 회원 1, 2, 3생성
         memberId1 = createMember("name1", "nickName1", "01011111111");
         memberId2 = createMember("name2", "nickName2", "01022222222");
@@ -375,7 +375,7 @@ class ReviewServiceTest {
     }
 
     public Long createBoard(Long savedId, String title, String content, LocalDateTime startTime, LocalDateTime endTime,
-                            TransactionAddress address, Long price, boolean isFreebie) {
+                            TransactionAddress address, Long price, boolean isFreebie) throws Exception {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd-HH:mm");
         String formatStartTime = startTime.format(formatter);
         String formatEndTime = endTime.format(formatter);
